Convert MessageBox PickOption parameter to object
The Avalonia Xaml generator always generates a call to TryConvert even if there is a direct type converter specified, and it's known at xaml compile time that the types are compatible.
This causes NativeAOT compilation to throw an IL trimming error, since TryConvert is marked as RequiresUnreferencedCode. The only case it doesn't is if the parameter type is `object`.

In this specific case, unless a user messes with the data template manually, we know that the PickOption parameter will always be of type MessageBoxResult.
This lets us safety change the parameter to object to avoid the IL trimming issue.

We're just doing a direct cast from object to MessageBoxResult. The existing generated code would result in an unbox exception if the type was anything other then an identical enum or integer value.
An integer value doesn't make sense, so we should be safe to just use MessageBoxResult directly.